refactor: シフトロジックをドメイン層へ移設
