+++ src/components/features/Shift/ShiftForm/pc/DailyView/DateRail.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Stack } from "@chakra-ui/react";
 import dayjs from "dayjs";
-import { getWeekdayLabel } from "../../utils/dateUtils";
+import { getWeekdayLabel } from "@/src/domains/shift/date";
 
 type Props = {
   dates: string[];
