+++ src/components/features/Shift/ShiftForm/pc/DailyView/ShiftPopover.tsx
@@ -2,9 +2,9 @@ import { Badge, Box, Flex, Portal, Text } from "@chakra-ui/react";
 import { useEffect } from "react";
 import { LuTrash2, LuX } from "react-icons/lu";
 import { IconButton } from "@/src/components/ui/Button";
+import { timeToMinutes } from "@/src/domains/shift/time";
+import type { ShiftData } from "@/src/domains/shift/types";
 import { BREAK_POSITION } from "../../constants";
-import type { ShiftData } from "../../types";
-import { timeToMinutes } from "../../utils/timeConversion";
 
 type ShiftPopoverProps = {
   shift: ShiftData | null;
