+++ src/components/features/Shift/ShiftForm/pc/DailyView/hooks/useDrag.ts
@@ -1,17 +1,17 @@
 import { useAtomValue, useSetAtom } from "jotai";
 import { useCallback, useRef, useState } from "react";
-import { DEFAULT_POSITION, RESIZE_EDGE_THRESHOLD } from "../../../constants";
-import { hourWidthAtom, selectedDateAtom, selectedPositionAtom, shiftConfigAtom, shiftsAtom } from "../../../stores";
-import type { DragMode, LinkedResizeTarget, ShiftData } from "../../../types";
 import {
-  detectLinkedResizeEdge,
   findShiftAtPosition,
   mergeAdjacentPositions,
   paintPosition,
   resizeLinkedPositions,
   resizePosition,
-} from "../../../utils/shiftOperations";
-import { pixelToMinutes } from "../../../utils/timeConversion";
+} from "@/src/domains/shift/operations";
+import type { DragMode, LinkedResizeTarget, ShiftData } from "@/src/domains/shift/types";
+import { DEFAULT_POSITION, RESIZE_EDGE_THRESHOLD } from "../../../constants";
+import { hourWidthAtom, selectedDateAtom, selectedPositionAtom, shiftConfigAtom, shiftsAtom } from "../../../stores";
+import { detectLinkedResizeEdge } from "../../../utils/hitTesting";
+import { pixelToMinutes } from "../../../utils/timelineGeometry";
 
 type DragState = {
   mode: DragMode;
