+++ src/components/features/Shift/ShiftForm/utils/timelineGeometry.ts
@@ -1,18 +1,5 @@
+import type { TimeRange } from "@/src/domains/shift/types";
 import { HOUR_WIDTH_PX, TIME_AXIS_PADDING_PX } from "../constants";
-import type { TimeRange } from "../types";
-
-// 時刻文字列 → 分（"10:30" → 630）
-export const timeToMinutes = (time: string): number => {
-  const [hours, minutes] = time.split(":").map(Number);
-  return hours * 60 + minutes;
-};
-
-// 分 → 時刻文字列（630 → "10:30"）
-export const minutesToTime = (totalMinutes: number): string => {
-  const hours = Math.floor(totalMinutes / 60);
-  const minutes = totalMinutes % 60;
-  return `${hours.toString().padStart(2, "0")}:${minutes.toString().padStart(2, "0")}`;
-};
 
 // X座標（ピクセル） → 分 + unitスナップ + timeRangeクランプ
 export const pixelToMinutes = (params: { x: number; timeRange: TimeRange; hourWidth?: number }): number => {
@@ -35,10 +22,4 @@ export const minutesToPixel = (minutes: number, timeRange: TimeRange, hourWidth:
 export const getTimeAxisWidth = (timeRange: TimeRange, hourWidth: number = HOUR_WIDTH_PX): number => {
   const totalHours = timeRange.end - timeRange.start;
   return TIME_AXIS_PADDING_PX * 2 + totalHours * hourWidth;
-};
-
-// 分を時間表示に変換（480 → "8h"）
-export const minutesToHoursLabel = (totalMinutes: number): string => {
-  const hours = Math.floor(totalMinutes / 60);
-  return `${hours}h`;
 };
