+++ src/components/features/ShiftBoard/DemoShiftBoardPage/DemoIntroTour.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, useEffect, useState } from "react";
-import type { ShiftData } from "@/src/components/features/Shift/ShiftForm/types";
 import { type EventData, Tour, type TourHandle, type TourStep } from "@/src/components/ui/Tour";
+import type { ShiftData } from "@/src/domains/shift/types";
 import { isPrerendering } from "@/src/helpers/seo";
 
 const PC_BREAKPOINT = 1024;
