+++ src/components/features/StaffSubmit/DayCard/index.tsx
@@ -1,8 +1,8 @@
 import { Box, Flex, Text } from "@chakra-ui/react";
 import { LuX } from "react-icons/lu";
-import { formatDateWithWeekday } from "@/src/components/features/Shift/ShiftForm/utils/dateUtils";
 import { IconButton } from "@/src/components/ui/Button";
 import { Select, type SelectItemType } from "@/src/components/ui/Select";
+import { formatDateWithWeekday } from "@/src/domains/shift/date";
 import { formatTime, getDateColor } from "../utils/timeOptions";
 
 export type DayEntry = {
