+++ src/components/features/StaffSubmit/ReadOnlySubmitView/index.tsx
@@ -1,8 +1,8 @@
 import { Box, Flex, Icon, Text, VStack } from "@chakra-ui/react";
 import { useMemo } from "react";
 import { LuInfo } from "react-icons/lu";
-import { getDateRange } from "@/src/components/features/Shift/ShiftForm/utils/dateUtils";
 import { STAFF_CONTENT_MAX_W } from "@/src/components/templates/StaffHeader";
+import { getDateRange } from "@/src/domains/shift/date";
 import { DayCard } from "../DayCard";
 import type { SubmissionData } from "../SubmitFormView";
 import { SubmitPageContent, SubmitPageHeader, SubmitPageLayout } from "../SubmitPageLayout";
