+++ src/domains/shift/time.ts
@@ -1,0 +1,18 @@
+// 時刻文字列 → 分（"10:30" → 630）
+export const timeToMinutes = (time: string): number => {
+  const [hours, minutes] = time.split(":").map(Number);
+  return hours * 60 + minutes;
+};
+
+// 分 → 時刻文字列（630 → "10:30"）
+export const minutesToTime = (totalMinutes: number): string => {
+  const hours = Math.floor(totalMinutes / 60);
+  const minutes = totalMinutes % 60;
+  return `${hours.toString().padStart(2, "0")}:${minutes.toString().padStart(2, "0")}`;
+};
+
+// 分を時間表示に変換（480 → "8h"）
+export const minutesToHoursLabel = (totalMinutes: number): string => {
+  const hours = Math.floor(totalMinutes / 60);
+  return `${hours}h`;
+};
